Add infrastructure code
Includes several CDI beans:

SqlProducer: gives out a groovy Sql object. The same Connection (as
tm_cz) is given out every time.
PropertiesProducer: produces Properties obejcts reads from the file
system. The filename is specified by adding the qualifer @FileRef to the
injection point.
PreliminaryStepObserver: cross bean steps are handled with events. I'm
not sure this is the best strategy, but that's what I opted for now. The
observer receives a StepExecution object and then generates a
"non-preliminary" event for the step if the StepExecution payload
doesn't specify that the step is to be skipped.